refactor: reorganize strapi components and types
- Moved components from strapi-sections/ to strapi/ directory for better organization
- Relocated types from types/strapi.ts to components/strapi/types.ts
- Updated all import references across the codebase
- Improved component organization and modularity
