+++ app/(core-products)/defi-wallet/page.tsx
@@ -3,10 +3,10 @@ import {notFound} from 'next/navigation';
 
 import {Button} from '@/components/common/Button';
 import {FooterBanner} from '@/components/FooterBanner';
-import {Card} from '@/components/strapi-sections/cards-row/Card';
-import CardsRow from '@/components/strapi-sections/cards-row/CardsRow';
+import {Card} from '@/components/strapi/cards-row/Card';
+import CardsRow from '@/components/strapi/cards-row/CardsRow';
 
-import type {TButton, TCard, TCardsRowSection, TFooterSection, TStrapiImage} from '@/types/strapi';
+import type {TButton, TCard, TCardsRowSection, TFooterSection, TStrapiImage} from '@/components/strapi/types';
 import type {ReactNode} from 'react';
 
 type TPage = {
