+++ app/(resources)/blog/(withNavigation)/tags/[tag]/ListOfPosts.tsx
@@ -9,7 +9,7 @@ import {IconChevron} from '@/components/common/icons/IconChevron';
 import {cl} from '@/components/utils/cl';
 import {useFetchPosts} from '@/hooks/useFetchPosts';
 
-import type {TBlogPost} from '@/types/strapi';
+import type {TBlogPost} from '@/components/strapi/types';
 import type {ReactElement} from 'react';
 
 const PAGE_SIZE = 12;
