+++ components/StrapiFAQ.tsx
@@ -5,7 +5,7 @@ import {useCallback, useEffect, useState} from 'react';
 import {QuestionSection} from '@/components/QuestionSection';
 import {getFaq} from '@/components/utils/query';
 
-import type {TFaqSectionItem} from '@/types/strapi';
+import type {TFaqSectionItem} from '@/components/strapi/types';
 import type {ReactNode} from 'react';
 
 export function StrapiFAQ(): ReactNode {
