+++ components/strapi/templates/ChainActions.tsx
