+++ components/strapi/templates/ChainHero.tsx
